+++ lib/picoscope_3000a_impl.cc
@@ -82,7 +82,7 @@ namespace gr {
     }
 
     static PS3000A_RANGE
-    convert_to_ps3000a_range(float range)
+    convert_to_ps3000a_range(double range)
     {
       if (range == 0.01)
         return PS3000A_10MV;
